creacion index

+++ inventarioVeterinariaMB/gestorProductos/views.py
@@ -7,6 +7,8 @@ from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 
 
+def es_superususario(user):
+    return user.is_superuser
 
 def index(request):
     return render(request, 'index.html')
@@ -62,7 +64,27 @@ def editarProducto(request, id):
     return render(request, 'productosTemplate/ingresarProductos.html', data)
 
 #funciones para categoria
+from django.http import HttpResponseRedirect
+from django.urls import reverse
+from django.shortcuts import render
+from django.contrib.auth.decorators import user_passes_test
+from .forms import CategoriaRegistroForm
+from django.http import HttpResponseRedirect
+from django.urls import reverse
+from django.shortcuts import render
+from django.contrib.auth.decorators import user_passes_test
+from django.contrib import messages
+from .forms import CategoriaRegistroForm
+
+def es_superusuario(user):
+    return user.is_superuser
+
+@user_passes_test(es_superusuario, login_url='home')
 def categoriaRegistro(request):
+    if not request.user.is_superuser:
+        
+        return HttpResponseRedirect(reverse('home'))
+
     form = CategoriaRegistroForm()
     
     if request.method == 'POST':
@@ -71,25 +93,35 @@ def categoriaRegistro(request):
             form.save()
             return HttpResponseRedirect(reverse('categoriasData'))
             
-    data = {'form':form} 
-    return render(request, 'ProductosTemplate/ingresarCategorias.html',data)
+    data = {'form': form}
+    return render(request, 'ProductosTemplate/ingresarCategorias.html', data)
 
 
-def eliminarCategoria(request,id):
-    categoria = Categoria.objects.get(id=id)
-    categoria.delete()
+def eliminarCategoria(request, id): 
+    if not request.user.is_superuser: 
+        messages.warning(request, 'No tienes permisos para eliminar una categoría. Debes ser un superusuario.') 
+        return HttpResponseRedirect(reverse('home')) 
+    categoria = Categoria.objects.get(id=id) 
+    categoria.delete() 
     return HttpResponseRedirect(reverse('categoriasData'))
 
+
+@user_passes_test(es_superusuario, login_url='home')
 def editarCategoria(request, id):
+    if not request.user.is_superuser:
+        return HttpResponseRedirect(reverse('home'))
     categoria = Categoria.objects.get(id=id)
+
     form = CategoriaRegistroForm(instance=categoria)
-    
+
     if request.method == 'POST':
         form = CategoriaRegistroForm(request.POST, instance=categoria)
         if form.is_valid():
             form.save()
             return HttpResponseRedirect(reverse('categoriasData'))
-            
-    data = {'form':form} 
+
+
+    data = {'form': form}
     return render(request, 'productosTemplate/ingresarCategorias.html', data)
+
